services/registry/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
PyJWT==2.9.0
cryptography==43.0.3
PyYAML==6.0.2
networkx==3.3
python-dotenv==1.0.1
